+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 #### Builtin
 
 - add ``REGEX_TIMEOUT`` variable for global timout milliseconds of regex methods
+- add ``String#escape`` method for escaping meta-characters of glob and regex patterns
 
 ### Changed